+++ lib/game/widget_builder.dart
@@ -231,11 +231,18 @@ class WidgetBuilder {
       ),
     );
 
-    if (_hasBasicGestureDetectors(game)) {
+    final hasBasicDetectors = _hasBasicGestureDetectors(game);
+    final hasAdvancedDetectors = _hasAdvancedGesturesDetectors(game);
+
+    if (hasBasicDetectors && hasAdvancedDetectors) {
+      print('WARNING: Both Basic and Advanced detectors detected. Advanced detectors will override basic detectors and the later will not receive events');
+    }
+
+    if (hasBasicDetectors) {
       widget = _applyBasicGesturesDetectors(game, widget);
     }
 
-    if (_hasAdvancedGesturesDetectors(game)) {
+    if (hasAdvancedDetectors) {
       widget = _applyAdvancedGesturesDetectors(game, widget);
     }